update race stream

+++ backend/internal/entity/race.go
@@ -9,10 +9,10 @@ type Race struct {
 }
 
 type RaceResult struct {
-	PlayerId     string `json:"player_id" db:"player_id"`
-	Position     int    `json:"position" db:"position"`
-	Distance     int64  `json:"distance" db:"distance"`
-	RaceTime     int64  `json:"race_time" db:"race_time"`
-	FinishedAt   int64  `json:"finish_time" db:"finish_time"`
-	CurrentSpeed int64  `json:"current_speed" db:"current_speed"`
+	PlayerId     string  `json:"player_id" db:"player_id"`
+	Position     int     `json:"position" db:"position"`
+	Distance     int64   `json:"distance" db:"distance"`
+	RaceTime     float64 `json:"race_time" db:"race_time"`
+	FinishedAt   int64   `json:"finish_time" db:"finish_time"`
+	CurrentSpeed int64   `json:"current_speed" db:"current_speed"`
 }
